requirements.txt
tk
python-vlc
